+++ tests/unit/test_struct_meta.py
